+++ src/Detector/BrowserDetector.php
@@ -2,7 +2,6 @@
 
 declare(strict_types=1);
 
-
 namespace Rilwanfit\YauaaPhp\Detector;
 
 use Rilwanfit\YauaaPhp\Contracts\DetectorInterface;
@@ -16,9 +15,11 @@ final class BrowserDetector implements DetectorInterface
         foreach ($this->patterns as $browser) {
             if (preg_match($browser['pattern'], $userAgent, $matches)) {
                 return [
-                    'type' => 'browser',
-                    'name' => $browser['name'],
-                    'version' => $matches['version'] ?? null,
+                    'agent' => [
+                        'type' => 'browser',
+                        'name' => $browser['name'],
+                        'version' => $matches['version'] ?? null,
+                    ]
                 ];
             }
         }