+++ src/Detector/UnknownDetector.php
@@ -2,7 +2,6 @@
 
 declare(strict_types=1);
 
-
 namespace Rilwanfit\YauaaPhp\Detector;
 
 use Rilwanfit\YauaaPhp\Contracts\DetectorInterface;
@@ -12,9 +11,11 @@ final class UnknownDetector implements DetectorInterface
     public function detect(string $userAgent): ?array
     {
         return [
-            'type' => 'unknown',
-            'name' => null,
-            'version' => null,
+            'agent' => [
+                'type' => 'unknown',
+                'name' => null,
+                'version' => null,
+            ],
         ];
     }
 }